requirements.txt
azure-ai-projects
azure-identity
azure-monitor-opentelemetry
opentelemetry-sdk
pydantic
pydantic-settings